Extract Helper, use for Struct generator
